rspec, clean

+++ app/services/mailgun_service.rb
@@ -1,29 +1,35 @@
+# frozen_string_literal: true
+
 require 'mailgun-ruby'
 require 'tempfile'
 
 class MailgunService
-  def self.send_component_email(to:, subject:, html_body:, pdf_data:, filename:)
+  def self.send_email(to:, subject:, html_body:, pdf_data: nil, filename: nil) # rubocop:disable Metrics/MethodLength
     api_key = Rails.application.credentials.dig(:mailgun, :api_key)
     domain  = Rails.application.credentials.dig(:mailgun, :domain)
-    mg_host = 'api.eu.mailgun.net'
+    mg_host = Rails.application.credentials.dig(:mailgun, :api_host)
     mg_client = Mailgun::Client.new(api_key, mg_host)
-
-    tempfile = Tempfile.new([File.basename(filename, '.pdf'), '.pdf'])
-    tempfile.binmode
-    tempfile.write(pdf_data)
-    tempfile.rewind
 
     message_params = {
       from: "TGT Aluminium <postmaster@#{domain}>",
       to: to,
       subject: subject,
-      html: html_body,
-      attachment: tempfile
+      html: html_body
     }
+
+    if pdf_data && filename
+      tempfile = Tempfile.new([File.basename(filename, '.pdf'), '.pdf'])
+      tempfile.binmode
+      tempfile.write(pdf_data)
+      tempfile.rewind
+      message_params[:attachment] = tempfile
+    end
 
     mg_client.send_message(domain, message_params)
   ensure
-    tempfile.close
-    tempfile.unlink
+    if tempfile
+      tempfile.close
+      tempfile.unlink
+    end
   end
 end
